Add Logging project to solution then consolidate getting assembly version
